+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [2.6.3]
+- Fix: update map limits using `moveOnlyMapArea` after camera zoom changes. [#267](https://github.com/RafaelBarbosatec/bonfire/issues/267)
+- Adds `Future updateDirectional(JoystickDirectional? directional)` in `Joystick`. [#269](https://github.com/RafaelBarbosatec/bonfire/issues/269)
+
 # [2.6.2]
 - Updated example.
 - Remove mandatory of the `SimpleDirectionAnimation` in  `SimpleAlly`, `SimpleEnemy`, `SimpleNpc` and `SimplePlayer`